requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
apipkg==1.5
atomicwrites==1.4.0
attrs==19.3.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
ddt==1.4.1
docopt==0.6.2
execnet==1.7.1
google-auth==1.15.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
kubernetes==11.0.0
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==6.0.1
pytest-assume==2.2.1
pytest-dependency==0.5.1
pytest-forked==1.1.3
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-rerunfailures==9.0
pytest-sugar==0.9.4
pytest-xdist==1.32.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
wcwidth==0.1.9
websocket-client==0.57.0
zipp==3.1.0
